requirements.txt
aiohttp==3.9.5
flask[async]==2.3.3
flask-cors==4.0.0
gunicorn==21.2.0
google-generativeai==0.7.1
google-api-python-client==2.135.0
google-auth-httplib2==0.2.0
python-dotenv==1.0.0
uvicorn
bytez